refactor: simplify managed OpenCode session wiring
